Remake

+++ src/main/java/com/accenture/flowershop/DAO/FlowerDAO.java
@@ -5,6 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
@@ -13,5 +14,6 @@ public interface FlowerDAO extends JpaRepository<FlowerEntity, Long> {
     FlowerEntity findFlowerByName(String name);
 
     @Query(value = "SELECT * FROM FLOWERS WHERE price>=?1 AND price<=?2", nativeQuery = true)
-    List<FlowerEntity> findFlowersByPriceRange(int priceFrom, int priceTo);
+    List<FlowerEntity> findFlowersByPriceRange(BigDecimal priceFrom, BigDecimal priceTo);
+
 }
